find_command: Allow .\ prefix for missing files
utils.find_command was inconsistent when given a missing
filename with an extension.

utils.which fails for .\\missing.cmd, which caused
utils.find_command to return None instead of determining
the appropriate command for the given input.

This was inconsistent because utils.which would return the
given input when it was an absolute path, allowing
utils.find_command to proceed and find the appropriate
command.